refacto: assemble components
